pyproject.toml
[project]
name = "xenrag"
version = "0.1.0"
description = "Enterprise-grade, research-oriented LangGraph application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph",
    "langchain-core",
    "pydantic",
    "python-dotenv",
    "typing-extensions",
    "langchain-ollama",
    "qdrant-client",
    "neo4j",
    "sentence-transformers",
    "langchain-huggingface",
    "rich",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/xenrag"]
